% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_mistnet_init}
\alias{cpp_mistnet_init}
\title{Initializes the mistnet shared library pointed to by the path}
\usage{
cpp_mistnet_init(path)
}
\arguments{
\item{path}{The shared library}
}
\description{
Initializes the mistnet shared library pointed to by the path
}
\keyword{internal}
