% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_vol2bird_get_wsr88d_site_location}
\alias{cpp_vol2bird_get_wsr88d_site_location}
\title{Returns the wsr88d site location file}
\usage{
cpp_vol2bird_get_wsr88d_site_location()
}
\value{
location of site location file
}
\description{
Returns the wsr88d site location file
}
\keyword{internal}
