% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_vol2bird_initialize}
\alias{cpp_vol2bird_initialize}
\title{Initializes the vol2birdR library}
\usage{
cpp_vol2bird_initialize()
}
\description{
Initializes the vol2birdR library
}
\keyword{internal}
