% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_vol2bird_namespace__store_main_thread_id}
\alias{cpp_vol2bird_namespace__store_main_thread_id}
\title{Sets the main thread id}
\usage{
cpp_vol2bird_namespace__store_main_thread_id()
}
\description{
Sets the main thread id
}
\keyword{internal}
