% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_vol2bird_set_wsr88d_site_location}
\alias{cpp_vol2bird_set_wsr88d_site_location}
\title{Sets the wsr88d site location file}
\usage{
cpp_vol2bird_set_wsr88d_site_location(loc)
}
\arguments{
\item{loc}{location of file}
}
\description{
Sets the wsr88d site location file
}
\keyword{internal}
