% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_vol2bird_version}
\alias{cpp_vol2bird_version}
\title{Initializes the mistnet shared library pointed to by the path}
\usage{
cpp_vol2bird_version()
}
\arguments{
\item{path}{The shared library}
}
\description{
Initializes the mistnet shared library pointed to by the path
}
\keyword{internal}
