% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{default_pytorch_version}
\alias{default_pytorch_version}
\title{The default 'LibTorch' version to install}
\usage{
default_pytorch_version()
}
\value{
a version string
}
\description{
Returns the newest 'LibTorch' version available for the current platform.
}
\keyword{internal}
