% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{get_install_urls}
\alias{get_install_urls}
\title{List of installation files to download}
\usage{
get_install_urls(
  version = default_pytorch_version(),
  type = install_type(version = version)
)
}
\arguments{
\item{version}{The 'LibTorch' version to install.}

\item{type}{The installation type for 'LibTorch'. Valid value is currently \code{"cpu"}.}
}
\value{
a named list with character urls
}
\description{
List the 'LibTorch' and 'MistNet' files to download as local files
in order to proceed with \code{\link[=install_mistnet_from_file]{install_mistnet_from_file()}}.
}
\seealso{
\itemize{
\item \code{\link[=install_mistnet_from_file]{install_mistnet_from_file()}}
}
}
