% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_exists}
\alias{install_exists}
\title{Checks if the torch and mistnet library has been installed or not.
Returns TRUE if both torch and mistnet libraries can be found, otherwise FALSE}
\usage{
install_exists()
}
\description{
Checks if the torch and mistnet library has been installed or not.
Returns TRUE if both torch and mistnet libraries can be found, otherwise FALSE
}
