% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_mistnet_nexrad}
\alias{install_mistnet_nexrad}
\title{Install Mistnet Nexrad file}
\usage{
install_mistnet_nexrad(
  reinstall = FALSE,
  path = install_path(),
  timeout = 1800,
  from_url = "http://mistnet.s3.amazonaws.com/mistnet_nexrad.pt",
  ...
)
}
\arguments{
\item{reinstall}{Re-install the model even if its already installed}

\item{path}{Optional path to install or check for an already existing installation.}

\item{timeout}{Optional timeout in seconds for large file download.}

\item{from_url}{From where the mistnet model file should be downloaded.}

\item{...}{other optional arguments (like \code{`load`} for manual installation).}
}
\description{
Installs the mistnet model file
}
\details{
Used to download and install the mistnet model file
}
