% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_path}
\alias{install_path}
\title{Returns the path of the mistnet libraries for specified version}
\usage{
install_path(version = "1.0")
}
\arguments{
\item{version}{The Mistnet version checked for}
}
\value{
the path to the libraries
}
\description{
Returns the path of the mistnet libraries for specified version
}
