% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{lib_installed}
\alias{lib_installed}
\title{Returns the path of the 'MistNet' libraries for specified version}
\usage{
lib_installed(library_name, install_path)
}
\arguments{
\item{library_name}{The name of the library searched for, either 'libmistnet' or 'LibTorch'}

\item{install_path}{The location where to look for the libraries}
}
\value{
if anything could be located or not
}
\description{
Returns the path of the 'MistNet' libraries for specified version
}
\keyword{internal}
