% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_volume.R
\name{load_volume}
\alias{load_volume}
\title{Loads a volume vol2bird-style meaning that a number of different file formats
are tried and eventually loaded.}
\usage{
load_volume(file)
}
\arguments{
\item{file}{name of file to be loaded}
}
\value{
The loaded volume as a Rcpp_PolarVolume that can be passed around
}
\description{
Loads a volume vol2bird-style meaning that a number of different file formats
are tried and eventually loaded.
}
