% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistnet_load.R
\name{mistnet_default}
\alias{mistnet_default}
\title{The default 'MistNet' version}
\usage{
mistnet_default()
}
\value{
the default 'MistNet' version
}
\description{
The default 'MistNet' version
}
\keyword{internal}
