% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{mistnet_exists}
\alias{mistnet_exists}
\title{Checks if the 'Torch' and 'MistNet' libraries have been installed or not.}
\usage{
mistnet_exists()
}
\value{
TRUE if both 'Torch' and 'MistNet' libraries can be found, otherwise FALSE
}
\description{
Checks if the 'Torch' and 'MistNet' libraries have been installed or not.
}
