% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{mistnet_install_lib}
\alias{mistnet_install_lib}
\title{Installs the library}
\usage{
mistnet_install_lib(
  library_name,
  library_url,
  install_path,
  source_path,
  filter,
  md5hash,
  inst_path
)
}
\arguments{
\item{library_name}{The name of the library searched for, either 'libmistnet' or 'libtorch'}

\item{library_url}{Where to fetch the library}

\item{install_path}{Where to put the library}

\item{source_path}{If library should be fetched from somewhere else}

\item{filter}{Not used}

\item{md5hash}{MD5 check}

\item{inst_path}{inst path}
}
\value{
if anything could be located or not
}
\description{
Installs the library
}
\keyword{internal}
