% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{mistnet_install_libs}
\alias{mistnet_install_libs}
\title{Installs the 'MistNet' libraries}
\usage{
mistnet_install_libs(version, type, install_path, install_config)
}
\arguments{
\item{version}{version to install}

\item{type}{what type of libraries to be installed}

\item{install_path}{Where libraries should be installed}

\item{install_config}{the library config}
}
\description{
Installs the 'MistNet' libraries
}
\keyword{internal}
