% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{mistnet_installed}
\alias{mistnet_installed}
\title{Checks if the 'LibTorch' and 'MistNet' libraries have been installed, and
that the mistnet model file has been downloaded.}
\usage{
mistnet_installed(path, verbose = FALSE)
}
\arguments{
\item{path}{Optional non-default file path to check for the mistnet model file.}

\item{verbose}{When TRUE print informative messages on missing library and model files.}
}
\value{
TRUE if the 'LibTorch' and 'MistNet' libraries can be found and the
and the MistNet model file can be located, otherwise FALSE.
}
\description{
Checks if the 'LibTorch' and 'MistNet' libraries have been installed, and
that the mistnet model file has been downloaded.
}
\seealso{
\itemize{
\item \link{mistnet_exists}
}
}
