% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistnet_load.R
\name{mistnet_start}
\alias{mistnet_start}
\title{Initialize the mistnet system if enabled.}
\usage{
mistnet_start(version = mistnet_default(), reload = FALSE)
}
\arguments{
\item{version}{version of mistnet library}

\item{reload}{if mistnet library should be reloaded or not, default FALSE.}
}
\description{
Initialize the mistnet system if enabled.
}
