% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PolarVolume-class}
\alias{PolarVolume-class}
\alias{Rcpp_PolarVolume-class}
\title{PolarVolume}
\description{
The polar volume object as defined by RAVE.

The Rcpp PolarVolume class
A polar volume
}
\keyword{internal}
