% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RaveIO-class}
\alias{RaveIO-class}
\alias{Rcpp_RaveIO-class}
\title{RaveIO routines}
\description{
Provides I/O routines using the rave framework

The Rcpp RaveIO class
Used when loading and saving objects
}
\keyword{internal}
