% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{skip_if_no_temp_access}
\alias{skip_if_no_temp_access}
\title{Skip test if tempdir() not accessible}
\usage{
skip_if_no_temp_access()
}
\description{
Some function tests require tempdir() to be writeable in package vol2birdR.
This helper function allows to skip a test if tempdir() is not accessible
Inspired by \url{https://testthat.r-lib.org/articles/skipping.html#helpers}.
}
\keyword{internal}
