% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Vol2Bird-class}
\alias{Vol2Bird-class}
\alias{Rcpp_Vol2Bird-class}
\title{Vol2Bird processor class}
\description{
The vol2bird processing class.
Provides methods for processing polar volumes/scans.

The Rcpp vol2bird processing class.
}
\examples{
processor<-Vol2Bird$new()
processor$process(c("/...../volume.h5"),config,"vp.h5","pvol.h5")
}
\seealso{
\code{\link[=vol2bird]{vol2bird()}}
}
\keyword{internal}
