% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Vol2BirdConfig-class}
\alias{Vol2BirdConfig-class}
\alias{Rcpp_Vol2BirdConfig-class}
\title{Vol2Bird configuration}
\description{
The vol2bird configuration used during processing

The Rcpp vol2bird configuration class.
Configuration instance used when processing data.
}
\examples{
# create a configuration object:
config<-Vol2BirdConfig$new()
# copy configuration object using clone method:
extraconfig<-config$clone()
# copy configuration object using copy constructor:
otherconfig<-Vol2BirdConfig$new(config)
}
\seealso{
\code{\link[=vol2bird_config]{vol2bird_config()}}
}
\keyword{internal}
