% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_config}
\alias{install_config}
\title{Contains a list of 'MistNet' libraries for the various OS's}
\usage{
install_config
}
\description{
Contains a list of 'MistNet' libraries for the various OS's
}
\keyword{internal}
