% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{mistnet_exists}
\alias{mistnet_exists}
\title{Checks if the torch and mistnet libraries have been installed or not.
Returns TRUE if both torch and mistnet libraries can be found, otherwise FALSE}
\usage{
mistnet_exists()
}
\description{
Checks if the torch and mistnet libraries have been installed or not.
Returns TRUE if both torch and mistnet libraries can be found, otherwise FALSE
}
