% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{mistnet_exists}
\alias{mistnet_exists}
\title{Checks if the 'LibTorch' and 'MistNet' libraries have been installed or not.}
\usage{
mistnet_exists()
}
\value{
TRUE if both 'LibTorch' and 'MistNet' libraries can be found, otherwise FALSE
}
\description{
Checks if the 'LibTorch' and 'MistNet' libraries have been installed or not.
}
\seealso{
\itemize{
\item \code{\link[=mistnet_installed]{mistnet_installed()}}
\item \code{\link[=install_mistnet]{install_mistnet()}}
}
}
