% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistnet_load.R
\name{mistnet_start}
\alias{mistnet_start}
\title{Initialize the 'MistNet' system if enabled.}
\usage{
mistnet_start(version = mistnet_default(), reload = FALSE)
}
\arguments{
\item{version}{version of 'MistNet' library}

\item{reload}{if 'MistNet' library should be reloaded or not, default FALSE.}
}
\description{
Initialize the 'MistNet' system if enabled.
}
\keyword{internal}
