% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsl2odim.R
\name{rsl2odim}
\alias{rsl2odim}
\title{Convert a NEXRAD polar volume file to an ODIM polar volume file}
\usage{
rsl2odim(file, pvolfile_out = "", config, verbose = TRUE)
}
\arguments{
\item{file}{Character (vector). Either a path to a single radar polar volume
(\code{pvol}) file containing multiple scans/sweeps, or multiple paths to scan
files containing a single scan/sweep. The file data format should be either 1)
\href{https://github.com/adokter/vol2bird/blob/master/doc/OPERA2014_O4_ODIM_H5-v2.2.pdf}{ODIM}
format, which is the implementation of the OPERA data information model in
the \href{https://support.hdfgroup.org/HDF5/}{HDF5} format, 2) NEXRAD format
supported by the \href{https://trmm-fc.gsfc.nasa.gov/trmm_gv/software/rsl/}{RSL library} or 3) Vaisala
IRIS (IRIS RAW) format.}

\item{pvolfile_out}{Character. File name. When provided, writes a polar
volume (\code{pvol}) file in the ODIM HDF5 format to disk. Useful for converting
RSL formats to ODIM, and for adding MistNet segmentation output.}

\item{config}{optional configuration object of class \code{Rcpp_Vol2BirdConfig},
typically output from \link{vol2bird_config}}

\item{verbose}{logical. When TRUE print profile output to console.}
}
\description{
Convert a NEXRAD polar volume file to an ODIM polar volume file
}
\seealso{
\itemize{
\item \code{\link[=vol2bird_config]{vol2bird_config()}}
}
}
