% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Vol2Bird}
\alias{Vol2Bird}
\title{Vol2Bird}
\description{
The vol2bird processing class.
Provides methods for processing polar volumes/scans. Typical usage can be
processor<-Vol2Bird$new()
processor$process(c("/...../volume.h5"),config,"vp.h5","pvol.h5")
}
