% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol2bird.R
\name{vol2bird}
\alias{vol2bird}
\title{Calculate a vertical profile (\code{vp}) from a polar volume (\code{pvol}) file}
\usage{
vol2bird(file, vpfile = "", pvolfile_out = "", config)
}
\arguments{
\item{file}{Character (vector). Either a path to a single radar polar volume
(\code{pvol}) file containing multiple scans/sweeps, or multiple paths to scan
files containing a single scan/sweep. Or a single \code{pvol} object. The file data format should be either 1)
\href{https://github.com/adokter/vol2bird/blob/master/doc/OPERA2014_O4_ODIM_H5-v2.2.pdf}{ODIM}
format, which is the implementation of the OPERA data information model in
the \href{https://support.hdfgroup.org/HDF5/}{HDF5} format, 2) a format
supported by the \href{https://trmm-fc.gsfc.nasa.gov/trmm_gv/software/rsl/}{RSL library} or 3) Vaisala
IRIS (IRIS RAW) format.}

\item{vpfile}{Character. File name. When provided, writes a vertical profile
file (\code{vpfile}) in the ODIM HDF5 format to disk.}

\item{pvolfile_out}{Character. File name. When provided, writes a polar
volume (\code{pvol}) file in the ODIM HDF5 format to disk. Useful for converting
RSL formats to ODIM, and for adding MistNet segmentation output.}

\item{config}{optional configuration object of class \code{Rcpp_Vol2BirdConfig},
typically output from \link{vol2bird_config}}
}
\description{
Calculates a vertical profile of biological scatterers (\code{vp}) from a polar
volume (\code{pvol}) file using the algorithm
\href{https://github.com/adokter/vol2bird/}{vol2bird} (Dokter et al.
2011 \doi{10.1098/rsif.2010.0116}).
}
\seealso{
\itemize{
\item \code{\link[=vol2bird_config]{vol2bird_config()}}
}
}
