% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol2bird_config.R
\name{vol2bird_config}
\alias{vol2bird_config}
\title{Create a vol2bird configuration object}
\usage{
vol2bird_config()
}
\value{
an object of class \code{Vol2BirdConfig}
}
\description{
Creates a vol2bird configuration object of class \code{Vol2BirdConfig}
}
\examples{
# create a configuration object
config <- vol2bird_config()
# list the the configuration elements:
config
# change the maximum range included in the profile generation to 40 km:
config$rangeMax <- 40000
}
\seealso{
\itemize{
\item \code{\link[=vol2bird]{vol2bird()}}
}
}
