% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Vol2BirdConfig-class}
\alias{Vol2BirdConfig-class}
\alias{Rcpp_Vol2BirdConfig-class}
\title{Vol2Bird configuration}
\description{
The vol2bird configuration used during processing

The Rcpp vol2bird configuration class.
Configuration instance used when processing data.
}
\seealso{
\code{\link[=vol2bird_config]{vol2bird_config()}}
}
\keyword{internal}
