% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol2birdR.R
\docType{package}
\name{vol2birdR-package}
\alias{vol2birdR}
\alias{vol2birdR-package}
\title{vol2birdR: R-Variant of Vol2Bird}
\description{
R-variant of the vol2bird software.
}
\details{
To get started, see:

\itemize{
\item Dokter et al. (2016) \doi{https://doi.org/10.1098/rsif.2010.0116}: a
paper describing the profiling algorithmp.
\item \href{https://adriaandokter.com/vol2bird}{vol2bird C code documentation}:
an overview of the algorithm structure.
}
}
\keyword{internal}
