% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol2birdR.R
\docType{package}
\name{vol2birdR-package}
\alias{vol2birdR}
\alias{vol2birdR-package}
\title{vol2birdR: Vertical Profiles of Biological Signals in Weather Radar Data}
\description{
'R' implementation of the 'vol2bird' software for generating vertical profiles of birds and other biological signals in weather radar data. See Dokter et al. (2011) \doi{10.1098/rsif.2010.0116} for a paper describing the methodology.
}
\details{
To get started, see:

\itemize{
\item Dokter et al. (2016) \doi{https://doi.org/10.1098/rsif.2010.0116}: a
paper describing the profiling algorithm.
\item \href{https://adriaandokter.com/vol2bird/}{'vol2bird' 'C' code documentation}:
an overview of the algorithm structure.
\item Lin T-Y et al. (2019) \doi{https://doi.org/10.1111/2041-210X.13280}: a
paper describing the 'MistNet' model for rain segmentation.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adokter/vol2birdR/}
  \item \url{https://adriaandokter.com/vol2bird/}
  \item Report bugs at \url{https://github.com/adokter/vol2birdR/issues}
}

}
\author{
\strong{Maintainer}: Adriaan M. Dokter \email{amd427@cornell.edu} (\href{https://orcid.org/0000-0001-6573-066X}{ORCID})

Authors:
\itemize{
  \item Anders Henja \email{anders.henja@gmail.com}
}

Other contributors:
\itemize{
  \item Jurriaan H. Spaaks [contributor]
  \item Lourens Veen [contributor]
  \item Iwan Holleman [contributor]
  \item Hidde Leijnse [contributor]
  \item John H. Merritt [contributor]
  \item Bart Kelley [contributor]
  \item Daniel Michelson [contributor]
  \item Mark Couture [contributor]
  \item GloBAM (https://globam.science) [funder]
}

}
\keyword{internal}
